backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
neo4j==5.16.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
openai==1.10.0
langchain-community==0.0.38
langchain-core==0.1.52
requests-toolbelt==1.0.0
ollama==0.1.7